Move the generation of the core tables to install
I wanted to merit multiple models. I'm not sure if this is going to
work in the long run, but the generators wouldn't let me run:

    rails g merit user
    rails g merit post

The merit generator was also creating the migrations for the core
tables, like badges and sashes. I moved that to the install
generator and left only the model specific stuff in merit.