Sash creation bug fix
When there's more than one rule pointing to the same action (let's say,
a badge and a points rule), and when a target user with no prior
reputation triggers both, two sashes are created, and part of the
reputation was lost.

Reloading the potentially just-created related sash before creating it
solves this issue.

[fixes #203] [fixes #204]