Initialise merit for API only apps when needed
Previously merit was initialised once when action_controller was loaded,
but that could happen either for base or for API, and if the app was
configured for the other one Merit wouldn't run.

After https://github.com/rails/rails/pull/28402 we can specifically hook
in to web or API, thus removing the need for the `run_once` option.

Thank you for your work on this, @jamesjefferies.

[fixes #297]