Remove Mongoid support
Merit hasn't been running automated tests on mongo id since:

```
commit c6ec066d52e0d852356c8727935a6961216280f3
Author: Tute Costa <tutecosta@gmail.com>
Date:   Sun Nov 9 14:27:10 2014 -0500
```

I don't use it in any deployment with this configuration so I can't
maintain it. I won't be able to test whether this code loads properly
with Zeitwerk, so removing for now.

If you need this code in your deployment, feel free to send a PR
reverting this commit and ensuring CI runs tests against it.

Also:

- Stops testing on unmaintained Rails 5.1.
- Drops unused JS files from dummy test app